requirements.txt
boto3==1.13.12
channels==2.4.0
channels-redis==2.4.2
daphne==2.5.0
Django==3.0.5
django-celery==3.3.1
django-friendship==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-storages==1.9.1
firebase-admin==4.3.0
gunicorn==19.9.0
Pillow==7.1.2
pyotp==2.3.0
PyJWT==1.7.1
psycopg2==2.8.5
psycopg2-binary==2.8.3
redis==2.10.6
requests==2.23.0
